[gui] move EFrameType from TGFrame.h to GuiTypes.h
Let use this enum in win32gdk to avoid direct dependency from
Gui library there
