+++ core/gui/inc/GuiTypes.h
@@ -373,4 +373,27 @@ enum ECursor { kBottomLeft, kBottomRight, kTopLeft, kTopRight,
                kMove, kCross, kArrowHor, kArrowVer, kHand, kRotate,
                kPointer, kArrowRight, kCaret, kWatch, kNoDrop };
 
+//---- types of frames (and borders), moved from TGFrame.h
+
+enum EFrameType {
+   kChildFrame      = 0,
+   kMainFrame       = BIT(0),
+   kVerticalFrame   = BIT(1),
+   kHorizontalFrame = BIT(2),
+   kSunkenFrame     = BIT(3),
+   kRaisedFrame     = BIT(4),
+   kDoubleBorder    = BIT(5),
+   kFitWidth        = BIT(6),
+   kFixedWidth      = BIT(7),
+   kFitHeight       = BIT(8),
+   kFixedHeight     = BIT(9),
+   kFixedSize       = (kFixedWidth | kFixedHeight),
+   kOwnBackground   = BIT(10),
+   kTransientFrame  = BIT(11),
+   kTempFrame       = BIT(12),
+   kMdiMainFrame    = BIT(13),
+   kMdiFrame        = BIT(14)
+};
+
+
 #endif
